Struct volatile_cell::VolatileCell

pub struct VolatileCell<T> {
    // some fields omitted
}

This structure is used to represent a hardware register. It is mostly used by the ioreg family of macros.

Methods

impl<T> VolatileCell<T>

const fn new(value: T) -> VolatileCell<T>

Create a cell with initial value.

fn get(&self) -> T

Get register value.

fn set(&self, value: T)

Set register value.